requirements.txt
boto3
paramiko
python-dotenv
loguru
